fix: added types
