+++ src/fold/base/__init__.py
@@ -1,7 +1,6 @@
 # Copyright (c) 2022 - Present Myalo UG (haftungbeschränkt) (Mark Aron Szulyovszky, Daniel Szemerey) <info@dreamfaster.ai>. All rights reserved. See LICENSE in root folder.
 
 from .classes import *  # noqa
-from .events import *
 from .utils import *
 
 feature_name_separator = "~"
